added location search functionality - added dependency injection code - also made POJO change for Float value of some params - added new Extension func.s

+++ app/src/main/java/com/amartya/weather/models/Current.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 data class Current(
     @SerializedName("last_updated_epoch" ) var lastUpdatedEpoch : Int?       = null,
     @SerializedName("last_updated"       ) var lastUpdated      : String?    = null,
-    @SerializedName("temp_c"             ) var tempC            : Int?       = null,
-    @SerializedName("temp_f"             ) var tempF            : Int?       = null,
+    @SerializedName("temp_c"             ) var tempC            : Double?       = null,
+    @SerializedName("temp_f"             ) var tempF            : Double?       = null,
     @SerializedName("is_day"             ) var isDay            : Int?       = null,
     @SerializedName("condition"          ) var condition        : Condition? = Condition(),
     @SerializedName("wind_mph"           ) var windMph          : Double?    = null,
